refactor TimeLimitingIterator
